Back to Physics
I had  originally decided to use non-physics movements to control my character.

This was a mistake as it does not allow me the benefits of using  the more advanced features of the physics engine (like penguin not flying through the floor lol). I have learnt the error of my ways and resorted to forces as movement with drag enhance controllability.