Some physics changes
A compromised setup that allows for playable if not necessarily realistic ground AND flying mode movements.

To get any better, I would need to engage different physics properties when in ground mode..

For example, air drag for rotation is minimal, but when on the ground, it is substantially more effort to turn. I am cheating using a zero friction terrain to get around this and keeping my gravity lighter / mass lower than it really should be  for accurate behavior on the ground.

I will have to decide if this is good enough !!??!!